xml/parsing/manakai/comments-1.dat

#data
<!--
< !--x-- >
< !--y-- >
--><a/>
#errors
1;1;no XML decl;;;s
2;6;parser:comment not closed;;;m
2;9;parser:comment not closed;;;m
3;6;parser:comment not closed;;;m
3;9;parser:comment not closed;;;m
#document escaped
| <!--\u0020
| < !--x-- >
| < !--y-- >
|  -->
| <{} a>

#data
<!--< !--x-- >< !--y-- >--><a/>
#errors
1;1;no XML decl;;;s
1;10;parser:comment not closed;;;m
1;13;parser:comment not closed;;;m
1;20;parser:comment not closed;;;m
1;23;parser:comment not closed;;;m
#document escaped
| <!--\u0020< !--x-- >< !--y-- > -->
| <{} a>

#data
<!-- xxx -- > yyy --><a/>
#errors
1;1;no XML decl;;;s
1;12;parser:comment not closed;;;m
#document
| <!--  xxx -- > yyy  -->
| <{} a>
